% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_ordering.R
\name{basic_ordering}
\alias{basic_ordering}
\title{Nearest-Neighbor Basic Ordering}
\usage{
basic_ordering(X, col)
}
\arguments{
\item{X}{a numeric matrix, containing numeric, or integer values; should not include primary ordering variable (i.e. time). These points should be replicated at each value of the primary ordering variable. Each column corresponds to a dimension of the space.}

\item{col}{an integer, specifying which column of X to order by (all other column orders are randomized).}
}
\value{
Returns an object of class \code{Xord}, which is a list containing the following elements:
  \item{X}{A matrix containing the original X matrix}
  \item{order}{A vector giving the ordering of observations}
}
\description{
Create gridded observation ordering based on ordering in one variable, and randomizing the rest.
}
