% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ordering.R
\name{bin_ordering}
\alias{bin_ordering}
\title{Nearest-Neighbor Binned Ordering}
\usage{
bin_ordering(X, bins, iters = 1000)
}
\arguments{
\item{X}{a numeric matrix, containing numeric, or integer values; should not include primary ordering variable (i.e. time). These points should be replicated at each value of the primary ordering variable. Each column corresponds to a dimension of the space.}

\item{bins}{an integer specifying the number of bins per dimension of X, or a numeric vector containing the number of bins to use for each dimension of X. If vector, length(bins) should be equal to dim(X)[2]. Number of bins in any dimension must be less than or equal to the number of unique values in that dimension.}

\item{iters}{Number of iterations of stochastic swapping to perform.}
}
\value{
Returns an object of class \code{Xord}, which is a list containing the following elements:
  \item{X}{A matrix containing the original X matrix}
  \item{order}{A vector giving the ordering of observations}
  \item{bins}{A matrix containing the bin assignments for each observation}
  \item{score}{Pooled variance of orderings across bins; used to judge which ordering is best when conducted multiple times}
}
\description{
Create gridded observation ordering based on multi-dimensional binning in non-primary ordering variable.
}
