% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_plot.R
\name{interp.plot}
\alias{interp.plot}
\title{Interpolating Image Plots}
\usage{
interp.plot(X, Z, bounds = c(0, 1), resolution = 100, save = FALSE, ...)
}
\arguments{
\item{bounds}{a vector of length 2 indicating the lower and upper bound for interpolation and plotting.}

\item{X}{a numeric matrix with 2 columns specifying the coordinates at the observed Z values.}

\item{Z}{a numeric vector of responses corresponding to the coordinates in X.}

\item{Zlim}{a numeric vector representing the lower and upper limit for Z to use in plotting.}

\item{save}{a logical, indicating whether to save the plot as a file, or output to the current graphical device.}

\item{...}{extra parameters, to be passed to the png function}

\item{resolution}{an integer, specifying the resolution to interpolate at (number of unique points in each dimension).}
}
\description{
Interpolate a potentially non-gridded surface and output as an interpolated raster plot.
}
