% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximin.R
\name{maximin}
\alias{maximin}
\title{Maximin Design}
\usage{
maximin(n = 100, m = 2, T = 10000, X.init = NULL)
}
\arguments{
\item{n}{an integer, specifying the number of samples to generate.}

\item{m}{an integer, specifying the number of dimensions.}

\item{T}{an integer, specifying the number of stochastic optimization iterations to perform.}

\item{X.init}{a matrix, representing an initial design. If not null, values of n and m are obtained from this initial design.}
}
\value{
Returns a list containing the following elements:
  \item{X}{A matrix, representing the optimized design}
  \item{prog}{A vector, containing the value of the criterion (minimum distance between points) over the \code{T} iterations}
  \item{time}{The amount of time taken to complete the optimization (in seconds)}
}
\description{
Create a stochastically optimized maximin design.
}
