% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtx_elem_const.R
\name{mtx_D_const}
\alias{mtx_D_const}
\title{Nearest-Neighbor Distance Matrix Construction}
\usage{
mtx_D_const(Xord, m)
}
\arguments{
\item{Xord}{an object of class \code{Xord}, returned by functions such as multistart_ordering or bin_ordering.}

\item{m}{an integer specifying the maximum number of neighbors to use for each point in \code{X}.}
}
\value{
Returns a numeric distance matrix with sparsity determined by the number of nearest-neighbors.
}
\description{
Creates a sparse numeric matrix containing pairwise distances between neighbors.
}
