% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtx_elem_const.R
\name{mtx_elem_const}
\alias{mtx_elem_const}
\title{Covariance Matrix Construction}
\usage{
mtx_elem_const(prim, Xord, nn, kernf, d, g)
}
\arguments{
\item{prim}{a numeric vector containing the primary ordering variable (i.e. time).}

\item{Xord}{an object of class \code{Xord}, returned by functions such as multistart_ordering or bin_ordering.}

\item{nn}{an integer specifying the maximum number of neighbors to use for each point in \code{X}.}

\item{kernf}{kernel function to use for kernel calculations when constructing.}

\item{...}{Further arguments to pass to the distance function, such as lengthscales and a nugget.}
}
\value{
Returns a numeric covariance matrix with sparsity determined by the number of nearest-neighbors, and non-zero elements calculated using the selected \code{distance} function.
}
\description{
Creates a sparse numeric matrix
}
