% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistart_ordering.R
\name{multistart_ordering}
\alias{multistart_ordering}
\title{Parallel Implementations of Ordering using a Multistart Scheme}
\usage{
multistart_ordering(X, type = c("random", "binned"), multistart = 10, ...)
}
\arguments{
\item{X}{a matrix, containing numeric, or integer values; should not include primary ordering variable (i.e. time). These points should be replicated at each value of the primary ordering variable. Each column corresponds to a dimension of the space.}

\item{type}{a string corresponding to which type of ordering to perform. Options inclode "random", which randomly orders each observation, and "binned", which attempts to uniformly order across dimensions using bins.}

\item{multistart}{an integer specifying the number of independent, parallelized iterations of ordering.}

\item{...}{optional arguments regarding parallelization.}
}
\value{
Returns an object of class \code{Xord}, which is a list containing the following elements:
  \item{X}{A matrix containing the original X matrix}
  \item{order}{A vector giving the ordering of observations}
  \item{bins}{A matrix containing the bin assignments for each observation}
  \item{score}{Pooled variance of orderings across bins; used to judge which ordering is best when conducted multiple times}
}
\description{
Performs a user-defined number of parallel implementations of an ordering scheme, and then selects the best design under the pooled-variance criterion
}
